requirements.txt
numpy
scipy
sklearn
pytorch
torchvision
matplotlib
tensorboard
wandb
pandas
plotly
tqdm